refactor DishListViewController to MVVM
